Move FileIO provision to Table
