+++ core/src/main/java/com/netflix/iceberg/BaseTransaction.java
@@ -23,6 +23,7 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.netflix.iceberg.exceptions.CommitFailedException;
+import com.netflix.iceberg.io.FileIO;
 import com.netflix.iceberg.util.Tasks;
 import java.util.List;
 import java.util.Map;
@@ -364,6 +365,11 @@ class BaseTransaction implements Transaction {
     @Override
     public Transaction newTransaction() {
       throw new UnsupportedOperationException("Cannot create a transaction within a transaction");
+    }
+
+    @Override
+    public FileIO io() {
+      return transactionOps.io();
     }
   }
 }
