+++ core/src/main/java/com/netflix/iceberg/TableOperations.java
@@ -19,7 +19,7 @@
 
 package com.netflix.iceberg;
 
-import com.netflix.iceberg.io.OutputFile;
+import com.netflix.iceberg.io.FileIO;
 
 /**
  * SPI interface to abstract table metadata access and updates.
@@ -55,7 +55,7 @@ public interface TableOperations {
   void commit(TableMetadata base, TableMetadata metadata);
 
   /**
-   * @return a {@link com.netflix.iceberg.FileIO} to read and write table data and metadata files
+   * @return a {@link FileIO} to read and write table data and metadata files
    */
   FileIO io();
